Solved the only 3 of 23 geometries showing problem
The issue was caused by the geometries being smaller than the raster cell size. I increased the buffering distance and now it works.

Now that works, I tidied up the failed experiments with geometry file types in the data folder so that the only PreZero landfill file is 'PZ_landfill_point4326.geojson'

The landfills.shp and Spanish-provinces.shp were converted to GeoJSONs in QGIS and the shapefiles removed.