Added a map function for methane raster and landfill points
The map displays the raster for a specified date over an area allowing the end user to see if the high emissions correlate with a landfill's location.

There is a problem at the moment with the multigeometry being passed to the timeseries function. At the moment it throws up an expeption of "‘OpenEO batch job failed: Exception during Spark execution: org.locationtech.jts.geom.TopologyException: found non-noded intersection between LINESTRING ( -2.3165024783487738 36.83407211344984, -2.316502478348773 36.83407211344984 ) and LINESTRING ( -2.3165024783487738 36.83407211344984, -2.3165024783487724 36.83407211344984 ) [ (-2.3165024783487738, 36.83407211344984, NaN) ]’}]"

When the named geometry is deleted, the problem moves to a different geometry. When only one geometry is loaded (such as the test landfill) then it functions as intended.