Starting work on Sentinel 2 data
It should be possible to extract a methane plume by using band math. To do this, Band 12 during a methane emission can be compared to Band 12 when there is no methane emission.

I tried to compare bands using two datacubes but currently band math of that sort is not supported. I will have to think of a workaround.

If this is successful it will give PreZero a 20m resolution dataset to locate any leaks at their landfills.