Added bounding option and save map feature
Added second csv file to the Sentinel 2 CH4 Multi-Band-Multi-Pass map containing bounding so now it can be large or small. The user can choose which file to use when loading the landfill file.

The Sentinel 2 CH4 Multi-Band-Multi-Pass map will now save to the data directory after the script has run.