Corrections on data processing and visualisation
Data processing:
The code now determines the difference between the median of B11 and B12 and adds that value to B12 to normalise the two bands. Then B12 is substituted from B11 as before.

Visualisation:
The V_min has been set to zero and the mask has a code that works out what data values are 1.5 standard deviations above the mean and applies a mask to all values below that number so that the plume can be seen on the map.