Completed work on Sentinel 2 Multiband Single Pass tool
This is now the tool that will be recommended as the multiband multi pass workbook was very complicated for the end user, even if it gave clearer results. I will place the MBMP in it's own folder for interested parties if they want to run a more detailed analysis.

Beginning work on updating bounding boxes to more tightly fit the landfills as per Dr Varon's advice.